We don't need more than 64 bits for timestamps. 64 bits gets us 292 billion years past the Unix epoch. Also, this is more consistent with ds-warp.

+++ src/matching_market.sol
@@ -37,7 +37,7 @@ contract MatchingMarket is MatchingEvents, ExpiringMarket {
         _;
     }
 
-    function MatchingMarket(uint lifetime) ExpiringMarket(lifetime, 0) {
+    function MatchingMarket(uint64 lifetime) ExpiringMarket(lifetime, 0) {
     }
 
     //return true if offers[low] priced less than or equal to offers[high]
